+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/), and this project
 adheres to [Semantic Versioning](https://semver.org/).
 
+## [Unreleased]
+
+[unreleased]: https://github.com/rogdham/bigxml/compare/v0.1.0...HEAD
+
+### :rocket: Added
+
+- Usual shortcuts to `HTTPAdapter.requests(method, ...)`: `.get`, `.options`, `.head`,
+  `.post`, `.put`, `.patch`, `.delete`
+
 ## [0.1.0] - 2020-10-31
 
 [0.1.0]: https://github.com/rogdham/sdkite/releases/tag/v0.1.0